update test script

+++ change_detection_s2looking.py
@@ -156,7 +156,7 @@ def main(hp, mode, weights, device, split, trained_path, saved_path, threshold=0
                     #               os.path.join(hungarian_cd_save_path, "{filename}.png".format(filename=filename)))
 
                     #cv2.imwrite(os.path.join(hungarian_cd_save_path, "cd_{filename}.png".format(filename=filename)), full_cm * 255)
-                    cm_img = cv2.resize(full_cm, (dataset.width, dataset.height), interpolation=cv2.INTER_NEAREST)
+                    cm_img = cv2.resize(full_cm, (dataset.width, dataset.height), interpolation=cv2.INTER_NEAREST).astype('int')
                     # cm_img = Image.fromarray(full_cm, mode='1')
                     # cm_img.save(os.path.join(hungarian_cd_save_path, "cd_{filename}.png".format(filename=filename)))
                     # cm_img = np.asarray(cm_img.resize((dataset.width, dataset.height)))*1
